Updates to tuned code
