+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/caching/SolversMotor.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.hardware.caching;
 
+import com.outoftheboxrobotics.photoncore.Photon;
 import com.outoftheboxrobotics.photoncore.hardware.motor.PhotonDcMotor;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
