+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/example/ExampleConstants.java
@@ -10,4 +10,6 @@ public class ExampleConstants {
     public static double RIGHT_SERVO_POS = 0.50;
 
     public static double CENTER_SERVO_POS = 0.50;
+
+    public static double CENTER_MOTOR_POWER = 0.0;
 }
