+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/tuning/motor.java
@@ -1,21 +1,22 @@
-package org.firstinspires.ftc.teamcode.tuning.servo;
+package org.firstinspires.ftc.teamcode.tuning;
 
 import static org.firstinspires.ftc.teamcode.subsystem.System.checkButton;
-import static org.firstinspires.ftc.teamcode.subsystem.System.round;
-import static org.firstinspires.ftc.teamcode.tuning.example.ExampleConstants.CENTER_SERVO_POS;
+import static org.firstinspires.ftc.teamcode.tuning.example.ExampleConstants.*;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+import com.outoftheboxrobotics.photoncore.Photon;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.Gamepad;
 
 import org.firstinspires.ftc.teamcode.tuning.example.ExampleHardware;
 
+@Photon
 @Config
 @TeleOp
-public class wristTester extends OpMode {
+public class motor extends OpMode {
     public static boolean USE_DASHBOARD = false;
     ExampleHardware exampleHardware = new ExampleHardware();
     Gamepad currentGamepad1 = new Gamepad();
@@ -26,29 +27,26 @@ public class wristTester extends OpMode {
 
         telemetry = new MultipleTelemetry(telemetry, FtcDashboard.getInstance().getTelemetry());
 
-        exampleHardware.centerServo.setPosition(CENTER_SERVO_POS);
+        exampleHardware.intakeMotor.setPower(0);
     }
 
     @Override
     public void loop() {
         if (USE_DASHBOARD){
-            exampleHardware.centerServo.setPosition(CENTER_SERVO_POS);
+            exampleHardware.intakeMotor.setPower(CENTER_MOTOR_POWER);
         } else if (gamepad1.dpad_up  && checkButton(currentGamepad1, "dpad_up")) {
-            CENTER_SERVO_POS += 0.01;
+            CENTER_MOTOR_POWER += 0.01;
         } else if (gamepad1.dpad_down && checkButton(currentGamepad1, "dpad_down")) {
-            CENTER_SERVO_POS -= 0.01;
+            CENTER_MOTOR_POWER -= 0.01;
         }
 
         if (gamepad1.square || gamepad1.triangle || gamepad1.circle || gamepad1.cross) {
-            exampleHardware.centerServo.setPosition(CENTER_SERVO_POS);
+            exampleHardware.intakeMotor.setPower(CENTER_MOTOR_POWER);
         }
-
-        CENTER_SERVO_POS = Math.max(Math.min(CENTER_SERVO_POS, 1), 0);
-
+        ;
         currentGamepad1.copy(gamepad1);
 
-        telemetry.addData("centerServo getPosition", round(exampleHardware.centerServo.getPosition(), 2));
-        telemetry.addData("centerServoPos",round(CENTER_SERVO_POS, 2));
+        telemetry.addData("centerMotor Power", exampleHardware.intakeMotor.getPower());
         telemetry.update();
     }
 }
